backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
anthropic==0.75.0
mcp==1.25.0